+++ app/models/pay/charge.rb
@@ -17,6 +17,7 @@ module Pay
 
     # Store the payment method kind (card, paypal, etc)
     store_accessor :data, :paddle_receipt_url
+    store_accessor :data, :stripe_receipt_url
     store_accessor :data, :stripe_account
 
     # Payment method attributes