+++ CHANGELOG.md
@@ -40,6 +40,7 @@ end
 * Stripe.max_network_retries is now set to 2 by default. - @excid3
   This adds idempotency keys automatically to each request so that they can be safely retried.
 * Stripe Subscriptons can now be paused and resumed - @excid3
+* Store `stripe_receipt_url` on Pay::Charge - @mguidetti
 * Replace `update_email!` with `update_customer!` - @excid3
 * Add options for `cancel_now!` to support `invoice_now` and `prorate` flags for Stripe - @excid3
 * Adds `add_payment_processor` to add a payment processor without making it the default - @excid3